Fixed a bug in auxiliary functions for creating resonator with qubits and added a dark Layout theme
Fixed a bug, since the meander function  now returns a list in the tworesonators and the oneresonator functions in the auxiliary file.
Also added a dark theme for the GDSLayout viewer which can be activated by using sample.watch(dark = True). This does not give your eyes cancer like the standard color theme.